Thread/Local: cast to size_t, not long
A "long" has only 32 bits on WIN64, and is smaller than a pointer.
The warning that was supposed to be fixed by the intermediate "long"
cast reoccurs.

+++ src/Thread/Local.hpp
@@ -32,6 +32,8 @@ Copyright_License {
 #include <windows.h>
 #endif
 
+#include <stddef.h>
+
 /**
  * This class provides an interface to Thread Local Storage (TLS).
  */
@@ -92,11 +94,11 @@ template<class T>
 class ThreadLocalObject : private ThreadLocal {
 public:
   const T Get() const {
-    return (T)(long)ThreadLocal::Get();
+    return (T)(size_t)ThreadLocal::Get();
   }
 
   void Set(const T value) {
-    ThreadLocal::Set((void *)(long)value);
+    ThreadLocal::Set((void *)(size_t)value);
   }
 
   operator T() const {
